Removed unnecessary changes
